+++ zix/bitset.c
@@ -103,7 +103,7 @@ zix_bitset_count_up_to_if(const ZixBitset* b, const ZixBitsetTally* t, size_t i)
 
 	if (extra) {
 		const ZixBitset mask = ~(~(ZixBitset)0 << extra);
-		count += __builtin_popcountl(b[full_elems] & mask);
+		count += (size_t)__builtin_popcountl(b[full_elems] & mask);
 	}
 
 	return count;
